Logo Problem fix

+++ app/Http/Controllers/Admin/HomePageManagement.php
@@ -82,7 +82,11 @@ class HomePageManagement extends Controller
            $unique_name = md5(time() . rand()).".". $img -> getClientOriginalExtension();
 
            $img -> move(public_path('media/home'), $unique_name);
-           unlink('public/media/home/'. $request -> old_file);
+
+           if ( file_exists( 'public/media/home/'. $request -> old_file ) ) {
+               unlink('public/media/home/'. $request -> old_file);
+           }
+           
 
         }else {
             echo $unique_name = $request -> old_file;
@@ -197,7 +201,10 @@ class HomePageManagement extends Controller
            $unique_name = md5(time() . rand()).".". $img -> getClientOriginalExtension();
 
            $img -> move(public_path('media/home'), $unique_name);
-           unlink('public/media/home/'. $request -> old_file);
+           
+           if ( file_exists( 'public/media/home/'. $request -> old_file ) ) {
+               unlink('public/media/home/'. $request -> old_file);
+           }
 
         }else {
             echo $unique_name = $request -> old_file;
